+++ src/game/player/elements/scene/Scene.tsx
@@ -88,6 +88,9 @@ export default function Scene(
                         }
                         onClick();
                         state.stage.next();
+                        setTimeout(() => {
+                            usingSkipRef.current = false;
+                        }, 0);
                     }}
                     useTypeEffect={!usingSkipRef.current}
                 />